ref: move EditFieldDecorator to root storybook util
putting it in __tests__ causes cra's jest to complain T_T
