+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/EditFieldDrawer/edit-fieldtype/EditHomeno/EditHomeno.stories.tsx
@@ -4,9 +4,7 @@ import { BasicField, HomenoFieldBase } from '~shared/types'
 
 import { createFormBuilderMocks } from '~/mocks/msw/handlers/admin-form'
 
-import { StoryRouter } from '~utils/storybook'
-
-import { EditFieldDrawerDecorator } from '../__tests__/utils/storybook'
+import { EditFieldDrawerDecorator, StoryRouter } from '~utils/storybook'
 
 import { EditHomeno } from './EditHomeno'
