--- frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/EditFieldDrawer/edit-fieldtype/__tests__/utils/storybook.tsx
@@ -1,22 +1,0 @@
-import { Box, useDisclosure } from '@chakra-ui/react'
-import { DecoratorFn } from '@storybook/react'
-
-import { BuilderAndDesignContext } from '~features/admin-form/create/builder-and-design/BuilderAndDesignContext'
-import { CreatePageSidebarProvider } from '~features/admin-form/create/common/CreatePageSidebarContext'
-
-export const EditFieldDrawerDecorator: DecoratorFn = (storyFn) => {
-  const deleteFieldModalDisclosure = useDisclosure()
-  return (
-    <Box maxW="33.25rem">
-      <CreatePageSidebarProvider>
-        <BuilderAndDesignContext.Provider
-          value={{
-            deleteFieldModalDisclosure,
-          }}
-        >
-          {storyFn()}
-        </BuilderAndDesignContext.Provider>
-      </CreatePageSidebarProvider>
-    </Box>
-  )
-}
